refactor(application): reorganize use cases, DTOs and improve product creation flow
- Rename CreateCompanyIn to CreateSupplierIn
- Replace SlugProductCategoryIn with productCategoryIds in CreateSupplierIn
- Move CreateSupplierUseCase and ListSuppliersByCategoryUseCase to usecase/party package
- Refactor CreateProductUseCase to use NullValidator and Product.create() factory
- Add attribute mapping in CreateProductUseCase
- Remove unused @Builder from CreateProductOut record
- Fix package declarations across moved use cases
