--- core/application/src/main/java/com/arka/dto/in/SlugProductCategoryIn.java
@@ -1,7 +1,0 @@
-package com.arka.dto.in;
-
-
-public record SlugProductCategoryIn(
-        String slug
-) {
-}
