+++ core/application/src/main/java/com/arka/usecase/product/ListProductCategoriesUseCase.java
@@ -1,4 +1,4 @@
-package com.arka.usecase;
+package com.arka.usecase.product;
 
 import com.arka.gateway.repository.product.ProductCategoryGateway;
 import com.arka.model.product.ProductCategory;
